api/common-utils/src/main/java/com/roche/connect/common/order/dto/StatusDto.java
package com.roche.connect.common.order.dto;

public class StatusDto {
	private String status;
	private String workflowStatus;
	private String workflowType;
	
	

	public String getWorkflowStatus() {
		return workflowStatus;
	}

	public void setWorkflowStatus(String workflowStatus) {
		this.workflowStatus = workflowStatus;
	}

	public String getWorkflowType() {
		return workflowType;
	}

	public void setWorkflowType(String workflowType) {
		this.workflowType = workflowType;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}
	

}
